feat(game): Cleanup of pickable and rotatable groups, fixed window resize throwing items across map, fixed rotation not being canceled by panning
