+++ packages/game/src/types/runtime/EntityInstanceProps.ts
@@ -1,0 +1,9 @@
+import type { Block } from "../Block"
+import type { Stack } from "../Stack"
+
+export type EntityInstanceProps = {
+    stack: Stack,
+    block: Block,
+    rotation: number
+    variant?: number
+}
